casts to fix warnings in MSVC x64, thanks Willis
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@6045 d7cf8633-e32d-0410-b094-e92efae38249

+++ cl_screen.c
@@ -405,7 +405,7 @@ static void SCR_DrawDownload(void)
 	if (!cls.qw_downloadname[0])
 		return;
 	dpsnprintf(temp, sizeof(temp), "Downloading %s ...  %3i%%\n", cls.qw_downloadname, cls.qw_downloadpercent);
-	len = strlen(temp);
+	len = (int)strlen(temp);
 	x = (vid_conwidth.integer - len*size) / 2;
 	y = vid_conheight.integer - size;
 	DrawQ_Fill(0, y, vid_conwidth.integer, size, 0, 0, 0, 0.5, 0);
